* code by black * inference on custom dataset

+++ common_blocks/utils.py
@@ -15,7 +15,7 @@ from sklearn.model_selection import GroupKFold
 
 def seed_torch(seed=42):
     random.seed(seed)
-    os.environ['PYTHONHASHSEED'] = str(seed)
+    os.environ["PYTHONHASHSEED"] = str(seed)
     np.random.seed(seed)
     torch.manual_seed(seed)
     torch.cuda.manual_seed(seed)
@@ -23,8 +23,8 @@ def seed_torch(seed=42):
 
 
 def qwk_metric(y_pred, y, detach=True):
-    y_pred = torch.round(y_pred).to('cpu').numpy().argmax(1)
-    y = y.to('cpu').numpy()
+    y_pred = torch.round(y_pred).to("cpu").numpy().argmax(1)
+    y = y.to("cpu").numpy()
     # k = torch.tensor(cohen_kappa_score(torch.round(y_pred), y, weights='quadratic'), device='cuda:0')
     # k[k != k] = 0
     # k[torch.isinf(k)] = 0
@@ -33,18 +33,19 @@ def qwk_metric(y_pred, y, detach=True):
 
 
 def create_folds(configs):
-    if os.path.isfile(configs['folds_path']):
-        return pd.read_csv(configs['folds_path'])
+    if os.path.isfile(configs["folds_path"]):
+        return pd.read_csv(configs["folds_path"])
 
-    folds = pd.read_csv(configs['train_csv'])
-    train_labels = folds[configs['target_col']].values
-    kf = GroupKFold(n_splits=configs['nfolds'])
+    folds = pd.read_csv(configs["train_csv"])
+    train_labels = folds[configs["target_col"]].values
+    kf = GroupKFold(n_splits=configs["nfolds"])
 
     for fold, (train_index, val_index) in enumerate(
-            kf.split(folds.values, train_labels, folds['realative_coordinates'])):
-        folds.loc[val_index, 'fold'] = int(fold)
-    folds['fold'] = folds['fold'].astype(int)
-    folds.to_csv(configs['folds_path'], index=None)
+        kf.split(folds.values, train_labels, folds["realative_coordinates"])
+    ):
+        folds.loc[val_index, "fold"] = int(fold)
+    folds["fold"] = folds["fold"].astype(int)
+    folds.to_csv(configs["folds_path"], index=None)
     return folds
 
 
@@ -79,14 +80,16 @@ class OptimizedRounder(object):
             else:
                 X_p[i] = 4
 
-        ll = cohen_kappa_score(y, X_p, weights='quadratic')
+        ll = cohen_kappa_score(y, X_p, weights="quadratic")
         return -ll
 
     def fit(self, X, y):
         loss_partial = partial(self._kappa_loss, X=X, y=y)
         initial_coef = [0.5, 1.5, 2.5, 3.5]
-        self.coef_ = sp.optimize.minimize(loss_partial, initial_coef, method='nelder-mead')
-        print(-loss_partial(self.coef_['x']))
+        self.coef_ = sp.optimize.minimize(
+            loss_partial, initial_coef, method="nelder-mead"
+        )
+        print(-loss_partial(self.coef_["x"]))
 
     def predict(self, X, coef):
         X_p = np.copy(X)
@@ -104,11 +107,13 @@ class OptimizedRounder(object):
         return X_p
 
     def coefficients(self):
-        return self.coef_['x']
+        return self.coef_["x"]
 
 
 def convert_model(model, full_checkpoint_path, output_path):
-    checkpoint = torch.load(full_checkpoint_path, map_location=lambda storage, loc: storage)
+    checkpoint = torch.load(
+        full_checkpoint_path, map_location=lambda storage, loc: storage
+    )
 
     sanitized_dict = {}
     for k, v in checkpoint["state_dict"].items():
@@ -118,7 +123,7 @@ def convert_model(model, full_checkpoint_path, output_path):
     sample = torch.rand(1, 3, 256, 256, dtype=torch.float32)
     model.load_state_dict(sanitized_dict)
     scripted_model = torch.jit.trace(model, sample)
-    filename = ntpath.basename(full_checkpoint_path).replace('=', '')
+    filename = ntpath.basename(full_checkpoint_path).replace("=", "")
     os.makedirs(output_path, exist_ok=True)
     scripted_model.save(f"{output_path}/{filename}.pth")
 
@@ -138,7 +143,5 @@ def load_obj(obj_path: str, default_obj_path: str = "") -> Any:
     obj_name = obj_path_list[0]
     module_obj = importlib.import_module(obj_path)
     if not hasattr(module_obj, obj_name):
-        raise AttributeError(
-            f"Object `{obj_name}` cannot be loaded from `{obj_path}`."
-        )
+        raise AttributeError(f"Object `{obj_name}` cannot be loaded from `{obj_path}`.")
     return getattr(module_obj, obj_name)
